+++ airframe-http/src/main/scala/wvlet/airframe/http/Http.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.rest
+package wvlet.airframe.http
 
 sealed abstract class HttpMethod(name: String)
 object HttpMethod {
