+++ airframe-http/src/main/scala/wvlet/airframe/http/Router.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.rest
+package wvlet.airframe.http
 
 import javax.ws.rs.{DELETE, GET, POST, PUT}
 import wvlet.surface
