+++ airframe-http/src/test/scala/wvlet/airframe/http/ServiceExample.scala
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package wvlet.airframe.rest
+package wvlet.airframe.http
 
 import java.util.UUID
